Tools Folder Necessary for MKR1000 WiFi 101 Library to Work
This is the zip file that you're going to want to add to your Arduino sketchbook (main folder for Arduino files - should be in User/Documents/Arduino). Unarchive the zip file here and proceed to have the folder present in this location. Now, when you run Firmware Updater, you should have the option to select the latest version (up to version 1.6).